+++ release.nix
@@ -1,14 +1,8 @@
-{ supportedSystems ? [ "x86_64-linux" ]
+{ supportedSystems ? [ "x86_64-linux" "aarch64-darwin" "x86_64-darwin" ]
 }:
 rec {
   recurseForDerivations = true;
   build = import ./all-builds.nix { inherit supportedSystems; };
   test = import ./all-tests.nix { inherit supportedSystems; };
-  beta = {
-    skeleton = (import ./skeleton { system = "aarch64-darwin"; }).shells.ghc;
-    skeletonExe = (import ./skeleton { system = "aarch64-darwin"; }).exe;
-    command = (import ./.{ system = "aarch64-darwin"; }).command;
-  };
-
   inherit (build) metaCache;
 }